Bug Fix: Fixed bug that prevented cards Gas Calculator Cards to not fade in page after 3rd click of calculate button
Fix was achieved by unbinding the click event from the button immediately after triggering each click event.  This prevented multiple click events from being bound to the recalculate button. which resulted in multiple callback function calls per click of recalc button. This in turn caused the isFaded variable to have it's value inverted twice. The resultant value of  isFaded becomes true which resuilts in the code in elementsFadeOut() not executing.