Sweet Alert functionality added
I used the sweetalert.js to add popups/alerts to historical fuel pages to encourage users on smaller devices to rotate their screen to see graphs better. I also used it to inform users of the accuracy of the calculated values; that is, if the user owned the vehicle for less than 60% of it's age the calculated values may not be very accurate.